rewrite the experiment for objects
the fn takes a mutable object and emits a new fn; continuations don't hold
the fn. the dispatcher promises to keep track of it for us.

the event queue is generic to the continuation type. this is temporary and
we expect to fold everything together with a cast, but for now, we're just
trying to thread the needle with types that match correctly and pass the
compiler...