cps, cps/environment: only desym environment access
This is a WIP commit.

Experiments shows that we only have to desym due to issues in
environment definition and access.

Upstream bug: https://github.com/nim-lang/Nim/issues/17851

This commit disable blanket desymification and only perform them
on environment as a workaround for the mentioned bug, which fixes
the foreign symbol tests.