cps: add a new rewrite for broken return AST
Not sure why the compiler produces this stuff, maybe its an artifact of
desugaring `return expr` into `result = expr; return`. Except its
`return result = expr` in this case.

However we cannot rewrite this into `result = expr; return`, because
the compiler wouldn't like that if the proc is nested within an another
proc that returns (due to closure capture analysis), even if you use the
`result` symbol within the return statement (the compiler is broken beyond
help).