cps: turn cpsJump into a rewrite pass
Instead of making cpsJump an annotation so that other scope
macros may direct the flow, we employ the use of a custom pragma
as an annotation to where the control-flow continuation should
be inserted.

This drastically simplifies the work that scope macros will have to
do, since all transformations would have been done before they reaches
the scope macros.