cps: make cpsMagic proc raises Defect when used outside {.cps.}
Currently we put a {.warning.} inside the proc which does... absolutely
nothing but generates an useless warning.

Options to catch this at compile-time are non-existant, unfortunately,
since the procs using these cpsMagic procedures must sem before
they can be processed by {.cps.}.

So instead we are gonna make these cpsMagic proc raises when the
"non-cps" version is used outside of {.cps.}. Should help us catches
bugs in the transform if we misses anything.