cps/spec: yet another take on xfrmDefer
Now we query whether there is any defer that will
cause the node to be rewritten.

We also preserve most of the structures that are
present in the original nodes.

And now xfrmDefer actually recurse into it's components.

Still ugly, but more workable now since we don't use
a shared state anymore.