cps: always wrap the while body in an stmtlist
The compiler tends to flatten an StmtList with one child, thus when
the body of `while` only have one statement that is a cps call,
saften operates on the children of the call instead of the call
itself.

We fix this by always wrapping the body of a while loop in an
StmtList.