cps/xfrm: prevent leakage of the exception being handled
While the continuation is not executing, we do not want anything
to touch the exception via the exception leaked via
setCurrentException().

This is likely inadequate and can't handle an exception being raised
in the exception handler, so we will need `finally` for this.

This commit also consume the exception right before a scope change.