cps/transform: build the manager at a later pass
The nature of the context manager mean that it can't compose with other
transform stages since it must run before the continuation runs and
after the continuation step completes. Thus only building the manager
at the very end works for us.

By having the guarantee that the manager will have full management right
over the continuation, we can now fix Nim's exception stack in the event
of an unhandled exception escaping from CPS (see the accompanying tests
for more details)