cps/transform: implement a context management system for exception (#175)
* cps/transform: implement a context management system for exception

Exception handlers in Nim operates on a stack built within the Exception
object. Thus to ensure that continuations don't corrupt the stack
outside of CPS, we generates a manager that configures the current exception
to the continuation's before running the continuation, then restore the
original on return.

Known issue:
- Clearing the exception in the continuation is not reimplemented yet
- We are still ignoring unhandled exceptions coming from the continuation

* cps/transform: implements the manager with genAst

It let us write AST in Nim :)

* cps/transform: build the manager at a later pass

The nature of the context manager mean that it can't compose with other
transform stages since it must run before the continuation runs and
after the continuation step completes. Thus only building the manager
at the very end works for us.

By having the guarantee that the manager will have full management right
over the continuation, we can now fix Nim's exception stack in the event
of an unhandled exception escaping from CPS (see the accompanying tests
for more details)

* cps/transform: grammar, cleanup and documents

* cps/transform: clear handler's exception on scope exit

Implements a macro that will rewrite the body of exception handlers
and its continuation so that any attempt to exit the scope will clear
the stashed exception.

This is implemented as a macro so we can use it in templated bodies.

* cps/transform: only write finally body as exception when used as a handler

This let us remove cpsExceptionScope and merge it into withException,
allowing for a saner API and faster finally legs too.

* cps/transform: add missing comment

* cps/transform: sanity fixes

- Copy bodies that we are gonna edit
- Turn proc -> func where applicable
- Less `var`

* cps/transform: undo some diffs

Mostly artifacts from earlier testing