cps/transform: correctly transform try statement with one cps jump
The try statement AST have the first child as the body, which can be
inlined by the compiler when there are only one statement within.

In cpsCall handling we assume that control flow continues with the
following nodes, but that is not the case for try statements since the
nodes following the cpsCall are separated execution branches.

We fix this by building putting the cpsCall inside a statement list so
it is separated from the other execution branches.