cps/exprs: implement assignment flattening
Very experimental.

The way assignment flattening works is by copying the destination location
to every trailing expression. This comes with several restriction to the
assignment:

- The destination must be side-effect safe so that it can be evaluated
  in a different order (Right-to-Left)
- The destination expression (if any) must not contain any calls

This lets us support a large amount of assignments that we might not have
to expand any further.

An error message telling the user to create simpler flatten-able assignment
when we can't be certain of the evaluation order is also provided.