cps/exprs: put child bodies requiring analysis into StmtList
While annotating bodies that require flattening like a var/let
declaration, we have to run an analysis pass on a children of
those nodes.

However since `annotate()` inspect the children of a node, if
we don't place these children in a container, they will not be
inspected and only their children would be inspected.

The easiest container to use for this is StmtList since it has
no special function.