cps/rewrites: move varargs[typed] unpacking back to normalizing
For the purpose of expression flattening, we moved the majority of
normalizing rewrites into workaround rewrites in order to keep as
much semantic information as possible.

However `varargs[typed]` is one of the nodes whose semantic information
give us absolutely nothing as we cannot replicate them, and only the
compiler can build them.

Luckily they can only be the sole parameter of a proc (as far as we
know of), so losing this data does not impair our ability to keep
the evaluation ordering