cps: implements a pass for flattening expression (#182)
* cps/transform: move isCpsCall/isCpsBlock to spec

These will be used by the expression flattener

* cps/rewrites: move rewrites that de-sem the AST to workaround rewrites

These rewrites are only required because the compiler whine when you
touch those nodes. Otherwise they are correct nodes with useful
semantic data that we want to use in our transformations.

* cps: initial implementation of expression flattening

This implementation knows how to rewrite a `var/let` section so that
all expressions containing a cps block are evaluated outside of the section.

Only the expression list syntax is supported.

* cps/exprs: add flattening for block statement/expr

* cps/exprs: support flattening if expression

* cps/exprs: add case expression flattening

* cps/exprs: remove one workaround

It's not needed here

* cps/exprs: support flattening try expression

* cps/exprs: fix hasCpsExpr for StmtList

* cps/exprs: add flattening support for if condition

* cps/exprs: add flattening support for case matching expression

* cps/exprs: implement while condition flattening

* cps/exprs: implement assignment flattening

Very experimental.

The way assignment flattening works is by copying the destination location
to every trailing expression. This comes with several restriction to the
assignment:

- The destination must be side-effect safe so that it can be evaluated
  in a different order (Right-to-Left)
- The destination expression (if any) must not contain any calls

This lets us support a large amount of assignments that we might not have
to expand any further.

An error message telling the user to create simpler flatten-able assignment
when we can't be certain of the evaluation order is also provided.

* cps/exprs: put child bodies requiring analysis into StmtList

While annotating bodies that require flattening like a var/let
declaration, we have to run an analysis pass on a children of
those nodes.

However since `annotate()` inspect the children of a node, if
we don't place these children in a container, they will not be
inspected and only their children would be inspected.

The easiest container to use for this is StmtList since it has
no special function.

* cps/exprs: implements flattening for conversions

Conversions are flattened by inlining the conversion into the
expression tails of complex expressions as there are no side
effects in doing so, saving us the need to produce temporaries
for them.

* cps/exprs: fix conversion rewrites

`debugAnnotation` wraps things into `nnkStmtList`, which erases their
type, so we gotta take the node out.

* cps/exprs: implements return/discard flattening

These are simple, we just need to copy `return`/`discard` to each
expression tails.

* cps/exprs: implement array construction flattening

This is a huge step for expression flattening as we can now analyze
whether an operand may need to be lifted given a Left-to-Right
evaluation order.

Of course, this is all very experimental, but I implemented the rules
conservative enough that it should not cause any trouble.

* cps/exprs: flatten object and tuple constructions

* cps/rewrites: move varargs[typed] unpacking back to normalizing

For the purpose of expression flattening, we moved the majority of
normalizing rewrites into workaround rewrites in order to keep as
much semantic information as possible.

However `varargs[typed]` is one of the nodes whose semantic information
give us absolutely nothing as we cannot replicate them, and only the
compiler can build them.

Luckily they can only be the sole parameter of a proc (as far as we
know of), so losing this data does not impair our ability to keep
the evaluation ordering

* cps/spec: pragma blocks can be cps blocks too

* cps/exprs: call flattening with and/or short circuit support

* cps/exprs: raise statement flattening

* cps: flurry of compiler bug reports

* cps/exprs: switch to set of mutable symbol kinds

There are more mutable symbols than immutable ones

* cps/exprs: use more ConstructNodes