add a pass to optimize environment conversion
In each CPS legs, we take a Continuation then convert it into Env
everytime we want to access a local. This is an extremely slow process
under arc/orc due to type checks.

Add a pass that convert the continuation symbol exactly once, with hopes
that it would speed things up for us.