emit an error on getting the address of a cps expression
Following my analysis (linked in code), it appears to me that addressing
depends on the subtle, but undefined behavior of how "access" operations
interacts.

The best route for this would be to block the user from doing this as
`addr` is a low-level operation where these interactions can drastically
affect how the program works.

Of course, the user can assign the result of an expression into a
variable, then obtain the address from it since that one is
well-defined.

On a semi-related note, we are purging `nnkAddr` from the list of
AccessNodes as this node can't be created from user code. It appears
to me that this node is only used internally after the `addr` magic
is resolved.