implements flattening for magic calls
Magic calls are treated differently by the compiler, where `var T`
parameters will not cause the AST at the callsite to become
HiddenAddr(<expr>), which cause issues when analyzed by CPS.

As a workaround, we analyze the magic symbol itself to figure out
the true type of the parameters.

This commit also lifted the restrictions on `addr` since it functions
using the same code path.