move path filtering to a CI job instead
Github's brilliant path filtering system is incompatible with
it's required checks system, so as a workaround we move to an userspace
version which won't have this issue.

A new 'All check passes' job is added to match against in required
checks since the feature doesn't work if the matrix isn't evaluated (ie.
when its skipped).