add support for converting a cps call result
Conversions are simple and side-effect free, so to support them in
assignments we only have to tweak the shim so that it look for the call
within the expression and replace it with the result.

For `discard`, we rewrite the statement into an assignment with a
temporary to simulate the effect since we cannot "skip" the conversion
due to their interactions with destructors.