remove workaround for nim-lang/Nim#18411
The issue was that no-return control flow in an except branch can mess
up the exception stack to the point of crashing the program, hence the
workaround was to exploit the "always return" property of a
continuation body.

Now that the issue is fixed, we can get rid of this workaround for a
more correct implementation.