prevent escape in finally after unhandled exceptions
Previously, following Nim, we only change the "normal progression" path
of finally after unhandled exception to re-raise. This makes it possible
for control-flow within finally block to break out of the current scope,
and avoid re-raising the unhandled exception.

However, I believe that this is unoptimal and rarely, if ever, the
programmer's intention, and that the unhandled exception must propagate
upwards regardless of scope exits to notify the caller of the error.

This commit implements a new behavior where all scope exits in the
finally handler for unhandled exception will be overriden with
re-raising the exception.

This new behavior is /not/ inline with how Nim behaves at the moment.

Thanks @zerbina for finding this issue in Nimskull, prompting this fix
for CPS.