cir: introduce a library for creating custom nodes
Creating a new node kind so far been a series of copy-pasting, which is
a pain as I intend to expand the number of custom nodes for future
passes.

CIR is a library intended to automate this process somewhat. We will
lose some stricter type checking (for now), in exchange for more
flexibility in expression (ie. checking multiple node types via sets).

Hidden in this commit is also fixes to some parts of analyze.