spec: use explicit conversion for trampolineIt on nimskull
Somehow `move` allowed the Continuation to be converted into its
environment type, which should not be possible given nim semantics.

nimskull doesn't like it as it should, so turn this into an explicit
conversion over there.

Unfortunately Nim pukes on the correct version of the code, so we stick
with the old one for that.