transform: rewrite result/return before annotate
Previously we unconditionally rewrite `return` with the assumption that
all return belongs to the procedure being rewritten. This does not hold
when nested procedures are used. This issue was exposed thanks to
https://github.com/nim-works/nimskull/pull/1503 breaking an implicit
dependency on the compiler's automatic result transformation for
implicit returns.

This commit separate return rewrite from annotate and merge it with the
result rewrite, as both require awareness of nested procedures.